API重构工作 #1

+++ src/main/java/top/r3944realms/superleadrope/client/renderer/state/SuperLeashRenderState.java
@@ -31,7 +31,7 @@ public record SuperLeashRenderState(
         float thickness,      // 线宽(根据张力变化)
         float swingAngle,     // 当前摆动角度(弧度)
         float swingSpeed,      // 摆动速度(弧度/tick)
-        float maxDistance,      // 最大距离
+        Double maxDistance,      // 最大距离
         boolean isFirstPerson,        // 是否是第一人称
         BlockPos belowBlockPos     // 持有者位置
 ) {
